Fix up spectral org, wavelets -> tf, add avg_func selection
