+++ neurodsp/tests/test_timefrequency_wavelets.py
@@ -1,16 +1,16 @@
-""".  """
+"""Tests for timefrequency estimations using wavelets."""
 
-from neurodsp.spectral.wavelet import *
+from neurodsp.timefrequency.wavelets import *
 
 ###################################################################################################
 ###################################################################################################
 
 def test_morlet_transform(tsig):
 
-    out = morlet_transform(tsig, [5, 10, 15], fs=500)
+    out = morlet_transform(tsig, 500, [5, 10, 15])
     assert True
 
 def test_morlet_convolve(tsig):
 
-    out = morlet_convolve(tsig, 10, fs=500)
+    out = morlet_convolve(tsig, 500, 10)
     assert True
